Support mulitple backends
This commit adds support for one additional backend based on HTTP::Tiny,
a minimal HTTP client written in pure Raku. It also adds the mechanism
to support more backends as they become available.

As part of this change, this also drops the requirement on a curl
binary and replaces it by adding a dependency on File::Which (used to
determine if curl is installed) and on HTTP::Tiny (used as a the
fallback client when curl is not available). As HTTP::Tiny is pure Raku,
it is guaranteed to be available wherever we can run HTTP::Tinyish.

The code still prefers to use curl when it is available. Note that this
differs from the Perl version, which prefers the pure-Perl HTTP::Tiny to
curl if the scheme is supported.